NeoTask alpha done

+++ neomodel/src/de/zib/gndms/neomodel/gorfx/NeoTask.java
@@ -13,6 +13,7 @@ import org.neo4j.graphdb.Node;
 import org.neo4j.graphdb.Relationship;
 import org.neo4j.graphdb.RelationshipType;
 import org.neo4j.graphdb.index.Index;
+import org.neo4j.index.impl.lucene.ValueContext;
 
 import java.io.Serializable;
 import java.util.Calendar;
@@ -27,30 +28,37 @@ import java.util.List;
  * To change this template use File | Settings | File Templates.
  */
 public class NeoTask extends NodeGridResource implements TimedGridResourceItf {
-    private static final String TASK_STATE_P = "TASK_STATE_P";
-    private static final String WID_P = "WID_P";
-    private static final String PERMISSION_INFO_P = "PERMISSION_INFO_P";
-    private static final String PROGRESS_P = "PROGRESS_P";
-    private static final String MAX_PROGRESS_P = "MAX_PROGRESS_P";
-    private static final String FAULT_STRING_P = "FAULT_STRING_P";
-    private static final String DONE_P = "DONE_P";
-    private static final String BROKEN_P = "BROKEN_P";
-    private static final String DESCRIPTION_P = "DESCRIPTION_P";
-    private static final String ORQ_P = "ORQ_P";
-    private static final String CONTRACT_P = "CONTRACT_P";
-    private static final String SERIALIZED_CREDENTIAL_P = "SERIALIZED_CREDENTIAL_P";
-    private static final String POST_MORTEM_P = "POST_MORTEM_P";
-    private static final String TERMINATION_TIME_IDX = "terminationTimeIdx";
-    private static final String TERMINATION_TIME_P = "TERMINATION_TIME_P";
+    public static final String TASK_STATE_P = "TASK_STATE_P";
+    public static final String WID_P = "WID_P";
+    public static final String PERMISSION_INFO_P = "PERMISSION_INFO_P";
+    public static final String PROGRESS_P = "PROGRESS_P";
+    public static final String MAX_PROGRESS_P = "MAX_PROGRESS_P";
+    public static final String FAULT_STRING_P = "FAULT_STRING_P";
+    public static final String DONE_P = "DONE_P";
+    public static final String BROKEN_P = "BROKEN_P";
+    public static final String DESCRIPTION_P = "DESCRIPTION_P";
+    public static final String ORQ_P = "ORQ_P";
+    public static final String CONTRACT_P = "CONTRACT_P";
+    public static final String SERIALIZED_CREDENTIAL_P = "SERIALIZED_CREDENTIAL_P";
+    public static final String POST_MORTEM_P = "POST_MORTEM_P";
+    public static final String TERMINATION_TIME_IDX = "terminationTimeIdx";
+    public static final String TASK_STATE_IDX = "taskStateIdx";
+    public static final String TERMINATION_TIME_P = "TERMINATION_TIME_P";
 
 
     public static enum TaskRelationships implements RelationshipType {
         OFFER_TYPE_REL,
-        SUB_TASK_TYPE_REL
+        PARENT_REL
     }
 
     public NeoTask(@NotNull NeoReprSession session, @NotNull String typeNick, @NotNull Node underlying) {
         super(session, typeNick, underlying);
+        if (getProperty(TASK_STATE_P, null) == null)   {
+            final @NotNull String newTaskStateName = TaskState.INITIALIZED.name();
+            setProperty(TASK_STATE_P, newTaskStateName);
+            final @NotNull Index<Node> typeNickIndex = getTypeNickIndex(TASK_STATE_IDX);
+            typeNickIndex.add(repr(), session().getGridName(), newTaskStateName);
+        }
     }
 
     @Override
@@ -68,8 +76,13 @@ public class NeoTask extends NodeGridResource implements TimedGridResourceItf {
         return Enum.valueOf(TaskState.class, (String) getProperty(TASK_STATE_P));
     }
 
-    public void setTaskState(@NotNull TaskState taskState) {
-        setProperty(TASK_STATE_P, taskState.name());
+    public void setTaskState(@NotNull TaskState newTaskState) {
+        final @NotNull String oldTaskStateName = getTaskState().name();
+        final @NotNull String newTaskStateName = newTaskState.name();
+        setProperty(TASK_STATE_P, newTaskStateName);
+        final @NotNull Index<Node> typeNickIndex = getTypeNickIndex(TASK_STATE_IDX);
+        typeNickIndex.remove(repr(), session().getGridName(), oldTaskStateName);
+        typeNickIndex.add(repr(), session().getGridName(), newTaskStateName);
     }
 
     public @NotNull byte[] getSerializedCredential() {
@@ -194,25 +207,68 @@ public class NeoTask extends NodeGridResource implements TimedGridResourceItf {
 
     public @NotNull Iterable<NeoTask> getSubTasks() {
         final List<NeoTask> subTasks = new LinkedList<NeoTask>();
-        for (Relationship rel : repr().getRelationships(TaskRelationships.SUB_TASK_TYPE_REL, Direction.OUTGOING))
-            subTasks.add( new NeoTask(getReprSession(), getTypeNick(), rel.getEndNode()) );
+        for (Relationship rel : repr().getRelationships(TaskRelationships.PARENT_REL, Direction.INCOMING))
+            subTasks.add( new NeoTask(getReprSession(), getTypeNick(), rel.getStartNode()) );
+        return subTasks;
+    }
+
+    public @NotNull Iterable<NeoTask> getSubTasks(final NeoOfferType ot) {
+        final List<NeoTask> subTasks = new LinkedList<NeoTask>();
+        for (Relationship rel : repr().getRelationships(TaskRelationships.PARENT_REL, Direction.INCOMING)) {
+            final NeoTask task        = new NeoTask(getReprSession(), getTypeNick(), rel.getStartNode());
+            final NeoOfferType taskOT = task.getOfferType();
+            if ((ot == null) ? taskOT == null : ot.equals(taskOT))
+                subTasks.add(task);
+        }
         return subTasks;
     }
 
     public @Nullable NeoTask getParent() {
-        final Relationship rel = repr().getSingleRelationship(TaskRelationships.SUB_TASK_TYPE_REL, Direction.INCOMING);
+        final Relationship rel = repr().getSingleRelationship(TaskRelationships.PARENT_REL, Direction.OUTGOING);
         if (rel == null)
             return null;
         else
             return new NeoTask(getReprSession(), getTypeNick(), rel.getEndNode());
     }
 
-    public void setParent(@Nullable NeoTask task) {
-        final Relationship rel = repr().getSingleRelationship(TaskRelationships.SUB_TASK_TYPE_REL, Direction.INCOMING);
+    public final boolean hasParent() {
+        return getParent() != null;
+    }
+
+    public final boolean isSubTask() {
+        return hasParent();
+    }
+
+    public boolean isRootTask() {
+        return getParent() == null;
+    }
+
+    public @NotNull NeoTask getRootTask() {
+        final @Nullable NeoTask parent = getParent();
+        if (parent == null) return this; else return parent.getRootTask();
+    }
+
+
+    public boolean isBelowTask(@NotNull NeoTask task) {
+        NeoTask parent = this;
+        while (parent != null) {
+            if (task.equals(parent))
+                return true;
+            else
+                parent = parent.getParent();
+
+        }
+        return false;
+    }
+
+    public void setParent(@Nullable NeoTask parent) {
+        if (parent != null && parent.isBelowTask(this))
+            throw new IllegalArgumentException("New parent may not be a parent-chain child of this parent");
+        final Relationship rel = repr().getSingleRelationship(TaskRelationships.PARENT_REL, Direction.OUTGOING);
         if (rel != null)
             rel.delete();
-        if (task != null)
-            repr().createRelationshipTo(task.repr(getReprSession()), TaskRelationships.SUB_TASK_TYPE_REL);
+        if (parent != null)
+            repr().createRelationshipTo(parent.repr(getReprSession()), TaskRelationships.PARENT_REL);
     }
 
     public @Nullable Calendar getTerminationTime() {
@@ -230,7 +286,8 @@ public class NeoTask extends NodeGridResource implements TimedGridResourceItf {
         final Calendar oldTerminationTime = getTerminationTime();
         if (oldTerminationTime != null) {
             final Index<Node> index = getTypeNickIndex(TERMINATION_TIME_IDX);
-            index.remove(repr(), session().getGridName(), oldTerminationTime.getTimeInMillis());
+            index.remove(repr(), session().getGridName(),
+                    new ValueContext( oldTerminationTime.getTimeInMillis() ).indexNumeric());
         }
         if (terminationTime == null){
             repr().removeProperty(TERMINATION_TIME_IDX);
@@ -238,7 +295,8 @@ public class NeoTask extends NodeGridResource implements TimedGridResourceItf {
         else {
             setProperty(TERMINATION_TIME_P, terminationTime.getTimeInMillis());
             final Index<Node> index = getTypeNickIndex(TERMINATION_TIME_IDX);
-            index.add(repr(), session().getGridName(), terminationTime.getTimeInMillis());
+            index.add(repr(), session().getGridName(),
+                    new ValueContext( terminationTime.getTimeInMillis() ).indexNumeric());
 
         }
     }
